requirements.txt
torch
matplotlib
streamlit
pillow
torchvision
opencv-python